requirements.txt
google-genai
flask
flask-cors
langchain-google-genai
gunicorn
firebase-admin
pymysql
pandas
openpyxl
